Crate devtools

An actor-based remote devtools server implementation. Only tested with nightly Firefox versions at time of writing. Largely based on reverse-engineering of Firefox chrome devtool logs and reading of code.

Modules

actor 🔒

actors 🔒

https://searchfox.org/mozilla-central/source/devtools/server/actors

id 🔒

network_handler 🔒

protocol 🔒

Low-level wire protocol implementation. Currently only supports JSON packets.

resource 🔒

Structs

ActorMsg 🔒

DevtoolsInstance 🔒

EmptyReplyMsg 🔒

StreamId 🔒

Enums

UniqueId 🔒

Functions

allow_devtools_client 🔒

debugger_value_to_json 🔒

https://searchfox.org/mozilla-central/source/devtools/server/actors/object/utils.js#148

handle_client 🔒

Process the input from a single devtools client until EOF.

start_server

Spin up a devtools server that listens for connections on the specified port.