devtools::actor

Trait Actor

pub(crate) trait Actor:
    Any
    + ActorAsAny
    + Send
    + Sync
    + MallocSizeOf {
    // Required method
    fn name(&self) -> &str;

    // Provided methods
    fn handle_message(
        &self,
        request: ClientRequest<'_, '_>,
        registry: &ActorRegistry,
        msg_type: &str,
        msg: &Map<String, Value>,
        stream_id: StreamId,
    ) -> Result<(), ActorError> { ... }
    fn cleanup(&self, _id: StreamId) { ... }
}

A common trait for all devtools actors that encompasses an immutable name and the ability to process messages that are directed to particular actors.

Required Methods

fn name(&self) -> &str

Provided Methods

fn handle_message( &self, request: ClientRequest<'_, '_>, registry: &ActorRegistry, msg_type: &str, msg: &Map<String, Value>, stream_id: StreamId, ) -> Result<(), ActorError>

fn cleanup(&self, _id: StreamId)