devtools::actors

Module browsing_context

Liberally derived from the Firefox JS implementation. Connection point for remote devtools that wish to investigate a particular Browsing Context’s contents. Supports dynamic attaching and detaching which control notifications of navigation, etc.

Structs

BrowsingContextActor 🔒

The browsing context actor encompasses all of the other supporting actors when debugging a web view. To this extent, it contains a watcher actor that helps when communicating with the host, as well as resource actors that each perform one debugging function.

BrowsingContextActorMsg

BrowsingContextTraits 🔒

FrameUpdateMsg 🔒

FrameUpdateReply 🔒

ListWorkersReply 🔒

TabNavigated 🔒

Enums

TargetType 🔒