devtools::actors::browsing_context

Struct BrowsingContextActor

pub(crate) struct BrowsingContextActor {

    name: String,
    pub browser_id: DevtoolsBrowserId,
    active_pipeline_id: AtomicRefCell<PipelineId>,
    active_outer_window_id: AtomicRefCell<DevtoolsOuterWindowId>,
    pub browsing_context_id: DevtoolsBrowsingContextId,
    accessibility_name: String,
    pub console_name: String,
    css_properties_name: String,
    pub inspector_name: String,
    pub page_info: AtomicRefCell<DevtoolsPageInfo>,
    reflow_name: String,
    pub style_sheets_name: String,
    pub thread_name: String,
    script_chans: AtomicRefCell<FxHashMap<PipelineId, GenericSender<DevtoolScriptControlMsg>>>,
    pub watcher_name: String,

}

The browsing context actor encompasses all of the other supporting actors when debugging a web view. To this extent, it contains a watcher actor that helps when communicating with the host, as well as resource actors that each perform one debugging function.

Fields
name: Stringbrowser_id: DevtoolsBrowserId

This corresponds to webview_id

active_pipeline_id: AtomicRefCell<PipelineId>active_outer_window_id: AtomicRefCell<DevtoolsOuterWindowId>browsing_context_id: DevtoolsBrowsingContextIdaccessibility_name: Stringconsole_name: Stringcss_properties_name: Stringinspector_name: Stringpage_info: AtomicRefCell<DevtoolsPageInfo>reflow_name: Stringstyle_sheets_name: Stringthread_name: Stringscript_chans: AtomicRefCell<FxHashMap<PipelineId, GenericSender<DevtoolScriptControlMsg>>>watcher_name: String
Implementations

impl BrowsingContextActor

pub(crate) fn register( registry: &ActorRegistry, console_name: String, browser_id: DevtoolsBrowserId, browsing_context_id: DevtoolsBrowsingContextId, page_info: DevtoolsPageInfo, pipeline_id: PipelineId, outer_window_id: DevtoolsOuterWindowId, script_sender: GenericSender<DevtoolScriptControlMsg>, ) -> Arc<Self>

pub(crate) fn handle_new_global( &self, pipeline: PipelineId, script_sender: GenericSender<DevtoolScriptControlMsg>, )

pub(crate) fn title_changed(&self, pipeline_id: PipelineId, title: String)

pub(crate) fn frame_update(&self, request: &mut ClientRequest<'_, '_>)

pub fn simulate_color_scheme(&self, theme: Theme) -> Result<(), ()>

pub(crate) fn pipeline_id(&self) -> PipelineId

pub(crate) fn title(&self) -> String

pub(crate) fn url(&self) -> String

pub(crate) fn update_pipeline( &self, pipeline_id: PipelineId, outer_window_id: DevtoolsOuterWindowId, page_info: DevtoolsPageInfo, )

pub(crate) fn outer_window_id(&self) -> DevtoolsOuterWindowId

pub(crate) fn script_chan(&self) -> GenericSender<DevtoolScriptControlMsg>

Returns the script sender for the active pipeline.

pub(crate) fn instruct_script_to_send_live_updates( &self, should_send_updates: bool, )