devtools::actors::inspector

Module accessibility

The Accessibility actor is responsible for the Accessibility tab in the DevTools page. Right now it is a placeholder for future functionality.

Structs

AccessibilityActor

AccessibilityTraits 🔒

AccessibleWalkerActor

ActorMsg 🔒

BootstrapReply 🔒

BootstrapState 🔒

GetSimulatorReply 🔒

GetTraitsReply 🔒

GetWalkerReply 🔒

SimulatorActor