devtools::actors::inspector

Module node

This actor represents one DOM node. It is created by the Walker actor when it is traversing the document tree.

Structs

AttrMsg 🔒

DevtoolsEventListenerInfo 🔒

GetEventListenerInfoReply 🔒

GetUniqueSelectorReply 🔒

GetXPathReply 🔒

NodeActor 🔒

NodeActorMsg 🔒

Constants

MAX_INLINE_LENGTH 🔒

The maximum length of a text node for it to appear as an inline child in the inspector.

TEXT_NODE 🔒

Text node type constant. This is defined again to avoid depending on script, where it is defined originally. See script::dom::bindings::codegen::Bindings::NodeBinding::NodeConstants.