devtools::actors::inspector

Module walker

The walker actor is responsible for traversing the DOM tree in various ways to create new nodes

Structs

ChildrenReply 🔒

DocumentElementReply 🔒

GetInnerOrOuterHTMLReply 🔒

GetLayoutInspectorReply 🔒

GetMutationsReply 🔒

GetOffsetParentReply 🔒

MutationMsg 🔒

NewMutationsNotification 🔒

QuerySelectorReply 🔒

WalkerActor 🔒

WalkerMsg 🔒

WatchRootNodeNotification 🔒

Enums

MutationVariant 🔒

Functions

find_child

Recursively searches for a child with the specified selector If it is found, returns a list with the child and all of its ancestors. TODO: Investigate how to cache this to some extent.