devtools::actors

Module root

Connection point for all new remote devtools interactions, providing lists of know actors that perform more specific actions (targets, addons, browser chrome, etc.)

Liberally derived from the Firefox JS implementation.

Structs

DescriptorTraits 🔒

GetProcessResponse 🔒

GetRootReply 🔒

GetTabReply 🔒

GlobalActors 🔒

ListAddonsReply 🔒

ListProcessesResponse 🔒

ListServiceWorkerRegistrationsReply 🔒

ListTabsReply 🔒

ListWorkersReply 🔒

ProtocolDescriptionReply 🔒

RootActor 🔒

RootActorMsg 🔒

RootTraits 🔒

ServiceWorkerInfo 🔒

ServiceWorkerRegistrationMsg 🔒

Types 🔒

Enums

AddonMsg 🔒