Module devtools::actors::timeline

Structs

Emitter

FramerateEmitterReply

HighResolutionStamp
HighResolutionStamp is struct that contains duration in milliseconds with accuracy to microsecond that shows how much time has passed since actor registry inited analog https://w3c.github.io/hr-time/#sec-DOMHighResTimeStamp

IsRecordingReply

MarkersEmitterReply

MemoryEmitterReply

StartReply

StopReply

TimelineActor

TimelineMarkerReply

Statics

DEFAULT_TIMELINE_DATA_PULL_TIMEOUT