Struct devtools::actors::timeline::HighResolutionStamp

pub struct HighResolutionStamp(f64);

HighResolutionStamp is struct that contains duration in milliseconds with accuracy to microsecond that shows how much time has passed since actor registry inited analog https://w3c.github.io/hr-time/#sec-DOMHighResTimeStamp

Methods

impl HighResolutionStamp

fn new(start_stamp: PreciseTime, time: PreciseTime) -> HighResolutionStamp

fn wrap(time: f64) -> HighResolutionStamp