devtools::actors

Module watcher

The watcher is the main entry point when debugging an element. Right now only web views are supported. It talks to the devtools remote and lists the capabilities of the inspected target, and it serves as a bridge for messages between actors.

Liberally derived from the Firefox JS implementation.

Modules

network_parent

target_configuration

Liberally derived from https://searchfox.org/mozilla-central/source/devtools/server/actors/target-configuration.js This actor manages the configuration flags that the devtools host can apply to the targets.

thread_configuration

Liberally derived from https://searchfox.org/mozilla-central/source/devtools/server/actors/thread-configuration.js This actor manages the configuration flags that the devtools host can apply to threads.

Structs

DocumentEvent 🔒

GetBlackboxingActorReply 🔒

GetBreakpointListActorReply 🔒

GetNetworkParentActorReply 🔒

GetParentBrowsingContextIDReply 🔒

GetTargetConfigurationActorReply 🔒

GetThreadConfigurationActorReply 🔒

SessionContext 🔒

Describes the debugged context. It informs the server of which objects can be debugged. https://searchfox.org/mozilla-central/source/devtools/server/actors/watcher/session-context.js

WatchTargetsReply 🔒

WatcherActor 🔒

WatcherActorMsg 🔒

WatcherTraits 🔒

WillNavigateMessage 🔒

Enums

SessionContextType

TargetActorMsg 🔒