devtools

Struct DevtoolsInstance

pub(crate) struct DevtoolsInstance {
    pub(crate) registry: Arc<ActorRegistry>,
    pub(crate) id_map: Arc<Mutex<IdMap>>,
    pub(crate) browsing_contexts: FxHashMap<BrowsingContextId, String>,
    pub(crate) sender: Sender<DevtoolsControlMsg>,
    pub(crate) receiver: Receiver<DevtoolsControlMsg>,
    pub(crate) pipelines: FxHashMap<PipelineId, BrowsingContextId>,
    pub(crate) actor_workers: FxHashMap<WorkerId, String>,
    pub(crate) actor_requests: HashMap<String, String>,
    pub(crate) connections: Arc<Mutex<FxHashMap<StreamId, DevtoolsConnection>>>,
    pub(crate) next_resource_id: u64,
}

Fields
registry: Arc<ActorRegistry>id_map: Arc<Mutex<IdMap>>browsing_contexts: FxHashMap<BrowsingContextId, String>sender: Sender<DevtoolsControlMsg>

This is handed to clients so they can notify the devtools instance when their connection closes.

receiver: Receiver<DevtoolsControlMsg>pipelines: FxHashMap<PipelineId, BrowsingContextId>actor_workers: FxHashMap<WorkerId, String>actor_requests: HashMap<String, String>connections: Arc<Mutex<FxHashMap<StreamId, DevtoolsConnection>>>

A map of active TCP connections to devtools clients.

Client threads remove their connection from here once they exit.

next_resource_id: u64
Implementations

impl DevtoolsInstance

pub(crate) fn create( sender: Sender<DevtoolsControlMsg>, receiver: Receiver<DevtoolsControlMsg>, embedder: EmbedderProxy, ) -> Option<Self>

pub(crate) fn run(self)

pub(crate) fn handle_framerate_tick(&self, actor_name: String, tick: f64)

pub(crate) fn handle_navigate( &mut self, browsing_context_id: BrowsingContextId, state: NavigationState, )

pub(crate) fn handle_new_global( &mut self, ids: (BrowsingContextId, PipelineId, Option<WorkerId>, WebViewId), script_sender: GenericSender<DevtoolScriptControlMsg>, page_info: DevtoolsPageInfo, )

pub(crate) fn handle_title_changed( &self, pipeline_id: PipelineId, title: String, )

pub(crate) fn handle_console_resource( &mut self, pipeline_id: PipelineId, worker_id: Option<WorkerId>, resource: ConsoleResource, )

pub(crate) fn handle_dom_mutation( &mut self, pipeline_id: PipelineId, dom_mutation: DomMutation, ) -> Result<(), ActorError>

pub(crate) fn handle_clear_console( &mut self, pipeline_id: PipelineId, worker_id: Option<WorkerId>, )

pub(crate) fn find_console_actor( &self, pipeline_id: PipelineId, worker_id: Option<WorkerId>, ) -> Option<String>

pub(crate) fn handle_network_event( &mut self, connections: Vec<DevtoolsConnection>, request_id: String, network_event: NetworkEvent, )

pub(crate) fn create_network_event_actor( &mut self, request_id: String, browsing_context_name: String, ) -> String

Create a new NetworkEventActor for a given request ID and browsing context name.

pub(crate) fn handle_create_source_actor( &mut self, script_sender: GenericSender<DevtoolScriptControlMsg>, pipeline_id: PipelineId, source_info: SourceInfo, )

pub(crate) fn handle_update_source_content( &mut self, pipeline_id: PipelineId, source_content: String, )

pub(crate) fn handle_debugger_pause( &mut self, pipeline_id: PipelineId, frame_offset: FrameOffset, pause_reason: PauseReason, )

pub(crate) fn handle_create_frame_actor( &mut self, result_sender: GenericSender<String>, pipeline_id: PipelineId, frame: FrameInfo, )

pub(crate) fn handle_create_object_actor( &mut self, result_sender: GenericSender<String>, value: DebuggerValue, )

pub(crate) fn handle_create_environment_actor( &mut self, result_sender: GenericSender<String>, environment_info: EnvironmentInfo, parent: Option<String>, actor: Option<String>, )