Crate devtools

An actor-based remote devtools server implementation. Only tested with nightly Firefox versions at time of writing. Largely based on reverse-engineering of Firefox chrome devtool logs and reading of code.

Modules

actor ğŸ”’

actors ğŸ”’

https://searchfox.org/mozilla-central/source/devtools/server/actors

id ğŸ”’

network_handler ğŸ”’

protocol ğŸ”’

Low-level wire protocol implementation. Currently only supports JSON packets.

resource ğŸ”’

Structs

ActorMsg

DevtoolsInstance ğŸ”’

EmptyReplyMsg

StreamId ğŸ”’

Enums

UniqueId ğŸ”’

Functions

allow_devtools_client ğŸ”’

handle_client ğŸ”’

Process the input from a single devtools client until EOF.

start_server

Spin up a devtools server that listens for connections on the specified port.