Crate devtools

An actor-based remote devtools server implementation. Only tested with nightly Firefox versions at time of writing. Largely based on reverse-engineering of Firefox chrome devtool logs and reading of code.

Modules

actor

actors
Corresponds to http://mxr.mozilla.org/mozilla-central/source/toolkit/devtools/server/actors/

protocol
Low-level wire protocol implementation. Currently only supports [JSON packets] (https://wiki.mozilla.org/Remote_Debugging_Protocol_Stream_Transport#JSON_Packets).

Structs

ConsoleAPICall

ConsoleMsg

EventTimingsUpdateMsg

NetworkEventMsg

NetworkEventUpdateMsg

PageErrorMsg

ResponseStartUpdateMsg

SecurityInfoUpdateMsg

Constants

_IMPL_SERIALIZE_FOR_ConsoleAPICall

_IMPL_SERIALIZE_FOR_ConsoleMsg

_IMPL_SERIALIZE_FOR_EventTimingsUpdateMsg

_IMPL_SERIALIZE_FOR_NetworkEventMsg

_IMPL_SERIALIZE_FOR_NetworkEventUpdateMsg

_IMPL_SERIALIZE_FOR_PageErrorMsg

_IMPL_SERIALIZE_FOR_ResponseStartUpdateMsg

_IMPL_SERIALIZE_FOR_SecurityInfoUpdateMsg

Functions

run_server

start_server
Spin up a devtools server that listens for connections on the specified port.