Crate devtools

An actor-based remote devtools server implementation. Only tested with nightly Firefox versions at time of writing. Largely based on reverse-engineering of Firefox chrome devtool logs and reading of code.

Modules

actor 🔒

actors 🔒

https://searchfox.org/mozilla-central/source/devtools/server/actors

protocol 🔒

Low-level wire protocol implementation. Currently only supports JSON packets.

Structs

EmptyReplyMsg

EventTimingsUpdateMsg 🔒

NetworkEventMsg 🔒

NetworkEventUpdateMsg 🔒

ResponseStartUpdateMsg 🔒

SecurityInfoUpdateMsg 🔒

StreamId 🔒

Enums

UniqueId 🔒

Functions

allow_devtools_client 🔒

run_server 🔒

start_server

Spin up a devtools server that listens for connections on the specified port.