Struct devtools::actor::ActorRegistry

pub struct ActorRegistry {
    actors: HashMap<String, Box<Actor + Send>>,
    new_actors: RefCell<Vec<Box<Actor + Send>>>,
    old_actors: RefCell<Vec<String>>,
    script_actors: RefCell<HashMap<String, String>>,
    shareable: Option<Arc<Mutex<ActorRegistry>>>,
    next: Cell<u32>,
    start_stamp: PreciseTime,
}

A list of known, owned actors.

Fields

actors

new_actors

old_actors

script_actors

shareable

next

start_stamp

Methods

impl ActorRegistry

fn new() -> ActorRegistry

Create an empty registry.

fn create_shareable(self) -> Arc<Mutex<ActorRegistry>>

Creating shareable registry

fn shareable(&self) -> Arc<Mutex<ActorRegistry>>

Get shareable registry through threads

fn start_stamp(&self) -> PreciseTime

Get start stamp when registry was started

fn register_script_actor(&self, script_id: String, actor: String)

fn script_to_actor(&self, script_id: String) -> String

fn script_actor_registered(&self, script_id: String) -> bool

fn actor_to_script(&self, actor: String) -> String

fn new_name(&self, prefix: &str) -> String

Create a unique name based on a monotonically increasing suffix

fn register(&mut self, actor: Box<Actor + Send>)

Add an actor to the registry of known actors that can receive messages.

fn register_later(&self, actor: Box<Actor + Send>)

fn find<'a, T: Any>(&'a self, name: &str) -> &'a T

Find an actor by registered name

fn find_mut<'a, T: Any>(&'a mut self, name: &str) -> &'a mut T

Find an actor by registered name

fn handle_message(&mut self, msg: &BTreeMap<String, Value>, stream: &mut TcpStream) -> Result<(), ()>

Attempt to process a message as directed by its to property. If the actor is not found or does not indicate that it knew how to process the message, ignore the failure.

fn drop_actor(&mut self, name: String)

fn drop_actor_later(&self, name: String)