Struct devtools::actor::ActorRegistry

pub struct ActorRegistry {
    actors: HashMap<String, Box<dyn Actor + Send>>,
    new_actors: RefCell<Vec<Box<dyn Actor + Send>>>,
    old_actors: RefCell<Vec<String>>,
    script_actors: RefCell<HashMap<String, String>>,
    shareable: Option<Arc<Mutex<ActorRegistry>>>,
    next: Cell<u32>,
    start_stamp: CrossProcessInstant,
}

A list of known, owned actors.

Fields
actors: HashMap<String, Box<dyn Actor + Send>>new_actors: RefCell<Vec<Box<dyn Actor + Send>>>old_actors: RefCell<Vec<String>>script_actors: RefCell<HashMap<String, String>>shareable: Option<Arc<Mutex<ActorRegistry>>>next: Cell<u32>start_stamp: CrossProcessInstant
Implementations

impl ActorRegistry

pub fn new() -> ActorRegistry

Create an empty registry.

pub(crate) fn cleanup(&self, id: StreamId)

pub fn create_shareable(self) -> Arc<Mutex<ActorRegistry>>

Creating shareable registry

pub fn shareable(&self) -> Arc<Mutex<ActorRegistry>>

Get shareable registry through threads

pub fn start_stamp(&self) -> CrossProcessInstant

Get start stamp when registry was started

pub fn register_script_actor(&self, script_id: String, actor: String)

pub fn script_to_actor(&self, script_id: String) -> String

pub fn script_actor_registered(&self, script_id: String) -> bool

pub fn actor_to_script(&self, actor: String) -> String

pub fn new_name(&self, prefix: &str) -> String

Create a unique name based on a monotonically increasing suffix

pub(crate) fn register(&mut self, actor: Box<dyn Actor + Send>)

Add an actor to the registry of known actors that can receive messages.

pub(crate) fn register_later(&self, actor: Box<dyn Actor + Send>)

pub fn find<'a, T: Any>(&'a self, name: &str) -> &'a T

Find an actor by registered name

pub fn find_mut<'a, T: Any>(&'a mut self, name: &str) -> &'a mut T

Find an actor by registered name

pub(crate) fn handle_message( &mut self, msg: &Map<String, Value>, stream: &mut TcpStream, id: StreamId, ) -> Result<(), ()>

Attempt to process a message as directed by its to property. If the actor is not found or does not indicate that it knew how to process the message, ignore the failure.

pub fn drop_actor(&mut self, name: String)

pub fn drop_actor_later(&self, name: String)