Trait devtools::actor::Actor

pub trait Actor: Any + ActorAsAny {
    fn handle_message(&self, registry: &ActorRegistry, msg_type: &str, msg: &BTreeMap<String, Value>, stream: &mut TcpStream) -> Result<ActorMessageStatus, ()>;
    fn name(&self) -> String;
}

A common trait for all devtools actors that encompasses an immutable name and the ability to process messages that are directed to particular actors. TODO: ensure the name is immutable

Required Methods

fn handle_message(&self, registry: &ActorRegistry, msg_type: &str, msg: &BTreeMap<String, Value>, stream: &mut TcpStream) -> Result<ActorMessageStatus, ()>

fn name(&self) -> String