Trait devtools::actor::Actor

pub(crate) trait Actor: Any + ActorAsAny {
    fn handle_message(
        &self, 
        registry: &ActorRegistry, 
        msg_type: &str, 
        msg: &Map<String, Value>, 
        stream: &mut TcpStream, 
        id: StreamId
    ) -> Result<ActorMessageStatus, ()>;

    fn name(&self) -> String;

    fn cleanup(&self, _id: StreamId) { ... }
}

A common trait for all devtools actors that encompasses an immutable name and the ability to process messages that are directed to particular actors. TODO: ensure the name is immutable

Required methods

fn handle_message(
&self,
registry: &ActorRegistry,
msg_type: &str,
msg: &Map<String, Value>,
stream: &mut TcpStream,
id: StreamId
) -> Result<ActorMessageStatus, ()>

fn name(&self) -> String

Loading content...
Provided methods

fn cleanup(&self, _id: StreamId)

Loading content...