Module devtools::actors

Corresponds to http://mxr.mozilla.org/mozilla-central/source/toolkit/devtools/server/actors/

Modules

console
Liberally derived from the Firefox JS implementation. Mediates interaction between the remote web console and equivalent functionality (object inspection, JS evaluation, autocompletion) in Servo.

framerate

inspector
Liberally derived from the Firefox JS implementation.

memory

network_event
Liberally derived from the Firefox JS implementation. Handles interaction with the remote web console on network events (HTTP requests, responses) in Servo.

object

performance

profiler

root

tab
Liberally derived from the Firefox JS implementation. Connection point for remote devtools that wish to investigate a particular tab's contents. Supports dynamic attaching and detaching which control notifications of navigation, etc.

thread

timeline

worker