Module devtools::actors

Corresponds to http://mxr.mozilla.org/mozilla-central/source/toolkit/devtools/server/actors/

Modules

browsing_context
Liberally derived from the [Firefox JS implementation] (http://mxr.mozilla.org/mozilla-central/source/toolkit/devtools/server/actors/webbrowser.js). Connection point for remote devtools that wish to investigate a particular Browsing Context's contents. Supports dynamic attaching and detaching which control notifications of navigation, etc.

console
Liberally derived from the [Firefox JS implementation] (http://mxr.mozilla.org/mozilla-central/source/toolkit/devtools/server/actors/webconsole.js). Mediates interaction between the remote web console and equivalent functionality (object inspection, JS evaluation, autocompletion) in Servo.

device

emulation

framerate

inspector
Liberally derived from the [Firefox JS implementation] (http://mxr.mozilla.org/mozilla-central/source/toolkit/devtools/server/actors/inspector.js).

memory

network_event
Liberally derived from the [Firefox JS implementation] (http://mxr.mozilla.org/mozilla-central/source/toolkit/devtools/server/actors/webconsole.js). Handles interaction with the remote web console on network events (HTTP requests, responses) in Servo.

object

performance

preference

process

profiler

root

stylesheets

tab

thread

timeline

worker