Module devtools::actors

https://searchfox.org/mozilla-central/source/devtools/server/actors

Modules

browsing_context

Liberally derived from the Firefox JS implementation. Connection point for remote devtools that wish to investigate a particular Browsing Context’s contents. Supports dynamic attaching and detaching which control notifications of navigation, etc.

console

Liberally derived from the Firefox JS implementation. Mediates interaction between the remote web console and equivalent functionality (object inspection, JS evaluation, autocompletion) in Servo.

device

framerate

inspector

Liberally derived from the Firefox JS implementation.

memory

network_event

Liberally derived from the Firefox JS implementation. Handles interaction with the remote web console on network events (HTTP requests, responses) in Servo.

object

performance

preference

process

Liberally derived from the Firefox JS implementation.

reflow

This actor is used for protocol purposes, it forwards the reflow events to clients.

root

Connection point for all new remote devtools interactions, providing lists of know actors that perform more specific actions (targets, addons, browser chrome, etc.)

stylesheets

tab

Descriptor actor that represents a web view. It can link a tab to the corresponding watcher actor to enable inspection.

thread

timeline

watcher

The watcher is the main entry point when debugging an element. Right now only web views are supported. It talks to the devtools remote and lists the capabilities of the inspected target, and it serves as a bridge for messages between actors.

worker