Module devtools::actors::inspector

Liberally derived from the Firefox JS implementation.

Modules

accessibility

The Accessibility actor is responsible for the Accessibility tab in the DevTools page. Right now it is a placeholder for future functionality.

css_properties

This actor holds a database of available css properties, their supported values and alternative names

highlighter

Handles highlighting selected DOM nodes in the inspector. At the moment it only replies and changes nothing on Servo’s side.

layout

The layout actor informs the DevTools client of the layout properties of the document, such as grids or flexboxes. It acts as a placeholder for now.

node

This actor represents one DOM node. It is created by the Walker actor when it is traversing the document tree.

page_style

The page style actor is responsible of informing the DevTools client of the different style properties applied, including the attributes and layout of each element.

style_rule

Liberally derived from https://searchfox.org/mozilla-central/source/devtools/server/actors/thread-configuration.js This actor represents one css rule group from a node, allowing the inspector to view it and change it. A group is either the html style attribute or one selector from one stylesheet.

walker

The walker actor is responsible for traversing the DOM tree in various ways to create new nodes

Structs

GetHighlighterReply 🔒

GetPageStyleReply 🔒

GetWalkerReply 🔒

InspectorActor

SupportsHighlightersReply 🔒