devtools::actors::inspector::walker

Function find_child

pub fn find_child(
    script_chan: &GenericSender<DevtoolScriptControlMsg>,
    pipeline: PipelineId,
    name: &str,
    registry: &ActorRegistry,
    node: &str,
    hierarchy: Vec<NodeActorMsg>,
    compare_fn: impl Fn(&NodeActorMsg) -> bool + Clone,
) -> Result<Vec<NodeActorMsg>, Vec<NodeActorMsg>>

Recursively searches for a child with the specified selector If it is found, returns a list with the child and all of its ancestors. TODO: Investigate how to cache this to some extent.