Module devtools::actors::root

Connection point for all new remote devtools interactions, providing lists of know actors that perform more specific actions (targets, addons, browser chrome, etc.)

Liberally derived from the Firefox JS implementation.

Structs

ActorTraits 🔒

DescriptorTraits

GetProcessResponse 🔒

GetRootReply 🔒

GetTabReply 🔒

ListAddonsReply 🔒

ListProcessesResponse 🔒

ListServiceWorkerRegistrationsReply 🔒

ListTabsReply 🔒

ListWorkersReply 🔒

ProtocolDescriptionReply

RootActor

RootActorMsg

Types

Enums

AddonMsg 🔒