Module devtools::actors::timeline

Structs

Emitter

FramerateEmitterReply

HighResolutionStamp
HighResolutionStamp is struct that contains duration in milliseconds with accuracy to microsecond that shows how much time has passed since actor registry inited analog https://w3c.github.io/hr-time/#sec-DOMHighResTimeStamp

IsRecordingReply

MarkersEmitterReply

MemoryEmitterReply

StartReply

StopReply

TimelineActor

TimelineMarkerReply

Constants

_IMPL_SERIALIZE_FOR_FramerateEmitterReply

_IMPL_SERIALIZE_FOR_IsRecordingReply

_IMPL_SERIALIZE_FOR_MarkersEmitterReply

_IMPL_SERIALIZE_FOR_MemoryEmitterReply

_IMPL_SERIALIZE_FOR_StartReply

_IMPL_SERIALIZE_FOR_StopReply

_IMPL_SERIALIZE_FOR_TimelineMarkerReply

Statics

DEFAULT_TIMELINE_DATA_PULL_TIMEOUT